refactor(calculator, racingcar): rename package
