+++ src/main/kotlin/racingcar/game/Application.kt
@@ -1,0 +1,7 @@
+package racingcar.game
+
+import racingcar.game.ui.CommandLineReceiver
+
+fun main() {
+    Game(CommandLineReceiver()).start()
+}
